# apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Mesh gateway
  version: 0.1.0
  summary: Trading fees become hourly AI inference credits, spent through an OpenAI-compatible gateway, served by Macs and frontier providers, and sold on a marketplace when unused.
  description: |
    Mesh is a token whose trading fees buy AI inference for the people holding it. A 1.5 % fee on every
    $MESH trade is swept once an hour; half goes to the network treasury and half is split, pro rata, across
    every wallet holding at least 1,000 tokens (time-weighted over the hour) as credits denominated in US
    dollars. A second engine, a share of the margin on paid requests and marketplace fees, pays into the same
    hourly pool (`GET /stats → usageShareEnabled`).
    Holders spend credits through this gateway with their own API keys: open models on Macs in the network at
    a flat price per million tokens, frontier models through zero-data-retention providers at list plus a 6 %
    markup that covers what the upstream charges Mesh (`GET /v1/models` carries both prices). Credits nobody
    will use are sold on the credit marketplace (`/market/*`, docs/MARKETPLACE.md), and anyone can buy credits
    from Mesh at face value with a prepaid balance (`/me/credits/buy`). A holder's first sign-in receives
    starter credits, which can be spent but not sold.

    What stands behind a credit, and what limits it:

    * **Reserve.** The holder share of every sweep is swapped to a stablecoin on chain and kept in a
...

...

...

...
servers:
...
security: []
paths:
...
  /nodes/{id}:
    get:
      tags: [Nodes]
      summary: Stats for one node
      description: Node token, or the session of the wallet that owns the node.
      operationId: getNode
      security: [{ nodeToken: [] }, { session: [] }]
      parameters: [{ $ref: '#/components/parameters/NodeId' }]
      responses:
        '200':
          description: Status, uptime, jobs, earnings and reputation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodeStats' }
              example:
                nodeId: node_7kd2pq9f
                wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k
                status: idle
                online: true
                busy: false
                chip: M2 Pro
                ramGb: 32
                loadAvg: 1.2
                agentVersion: 0.1.0
                models: ['llama3.1:8b']
                uptimePct24h: 97.4
                jobs24h: 140
                jobsDone24h: 138
                jobsFailed24h: 2
                tokens24h: 612000
                earnedUsd24h: 0.0367
                earnedUsdTotal: 0.412
                reputation: { window: 100, jobs: 100, successRate: 0.98, avgFirstTokenMs: 640, eligible: true, minSuccessRate: 0.8 }
                lastSeen: 1759487010
                createdAt: 1759300000
                offlineAfterSec: 60
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 7-day session from `POST /auth/verify`.
    apiKey:
      type: http
      scheme: bearer
      description: '`mesh_sk_…` from `POST /keys`. Works wherever an OpenAI key works.'
    nodeToken:
      type: http
      scheme: bearer
      description: '`mesh_node_…` from `POST /nodes/register`.'
  parameters:
...
    NodeId:
      name: id
      in: path
      required: true
      schema: { type: string, example: node_7kd2pq9f }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: unauthorized, message: valid session token required }
    NotFound:
      description: No such resource for this wallet.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: not_found }
...
  schemas:
    UnixSeconds:
      type: integer
      description: Unix time in seconds (UTC).
    Error:
      type: object
      description: Error shape outside `/v1`.
      required: [error]
      properties:
        error: { type: string, description: Machine-readable code. }
        message: { type: string }
        statusCode: { type: integer }
        requestId: { type: string }
        issues: { type: array, items: { type: object } }
...
    NodeStats:
      type: object
      properties:
        nodeId: { type: string }
        wallet: { type: string }
        status: { type: string, enum: [offline, busy, idle] }
        online: { type: boolean }
        busy: { type: boolean }
        chip: { type: [string, 'null'] }
        ramGb: { type: [number, 'null'] }
        loadAvg: { type: [number, 'null'] }
        agentVersion: { type: [string, 'null'] }
        models: { type: array, items: { type: string } }
        uptimePct24h: { type: number }
        jobs24h: { type: integer }
        jobsDone24h: { type: integer }
        jobsFailed24h: { type: integer }
        tokens24h: { type: integer }
        earnedUsd24h: { type: number }
        earnedUsdTotal: { type: number }
        reputation:
          type: object
          properties:
            window: { type: integer }
            jobs: { type: integer }
            successRate: { type: number }
            avgFirstTokenMs: { type: [number, 'null'] }
            eligible: { type: boolean }
            minSuccessRate: { type: number }
        lastSeen: { $ref: '#/components/schemas/UnixSeconds' }
        createdAt: { $ref: '#/components/schemas/UnixSeconds' }
        offlineAfterSec: { type: integer }
        pledge: { $ref: '#/components/schemas/NodePledge' }
    NodePledge:
      type: object
      description: Operator pledge status and whether the node may serve `trusted` jobs (docs/PRIVACY.md §5).
      properties:
        signed: { type: boolean }
        signedAt: { oneOf: [{ $ref: '#/components/schemas/UnixSeconds' }, { type: 'null' }] }
        chain: { type: [string, 'null'] }
        trusted: { type: boolean }
        trustedVia: { type: [string, 'null'], enum: [allowlist, stake+pledge, null] }
        allowlisted: { type: boolean }
        requiredStakeTier: { type: [string, 'null'], description: Stake tier name needed together with the pledge. }
        stakeTier: { type: [string, 'null'] }
        stakeOk: { type: boolean }
...